+++ spec/spec_config.rb
@@ -160,6 +160,10 @@ RSpec.configure do |config|
       Pakyow::Assets::Scripts::Terser.destroy
     end
 
+    if defined?(Pakyow::Loader)
+      Pakyow::Loader.reset
+    end
+
     remove_constants(
       (Object.constants - $original_constants).select { |constant_name|
         constant_name.to_s.start_with?("Test")